# swagger_documentation.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Backend Challenge : Surveys management api",
    "description": "Api to create, take and manage surveys, it operates on 5 main routes: \n\n- Surveys (create, update, delete, take, metrics/statistics)\n- Questions (create, update, delete, upload description image, list, metrics, statistics)\n- Entries (results when taking a survey - same previous operations)\n- Answers (same operations)\n- Auth (admin only) (register, login, logout, forgot password, reset password + manipulation operations on all objects)",
    "contact": {},
    "version": "1.0.2",
    "x-codegen-settings": {
      "Nullify404": false,
      "GenerateAsyncCode": true,
      "UseMethodPrefix": false,
      "UseModelPostfix": false,
      "UseControllerPrefix": true,
      "UseEnumPostfix": true,
      "CollectParameters": false,
      "UseConstructorsForConfig": false,
      "UseCommonSDKLibrary": false,
      "iOSUseAppInfoPlist": false,
      "AndroidUseAppManifest": false,
      "BodySerialization": 0,
      "EnableAdditionalModelProperties": false,
      "PreserveParameterOrder": true,
      "AppendContentHeaders": true,
      "iOSGenerateCoreData": false,
      "GenerateInterfaces": false,
      "NodeHttpClient": "NODE_REQUEST",
      "ValidateRequiredParameters": false,
      "JavaUsePropertiesConfig": false,
      "Timeout": 0,
      "StoreTimezoneInformation": false,
      "EnablePHPComposerVersionString": false,
      "EnableLogging": false,
...
      ],
...
  "servers": [
...
  "paths": {
...
    "/surveys/{id}": {
      "get": {
        "tags": [
          "Surveys"
        ],
        "summary": "Get survey by id",
        "description": "### Get survey by id\nreturn an object survey if found otherwise an empty object\n\n**protect route** : no\n**parameters** : id",
        "operationId": "Getsurveybyid",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "example": "null"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "X-Powered-By": {
                "style": "simple",
                "explode": false,
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "Express"
                  }
                }
              },
              "Content-Length": {
                "style": "simple",
                "explode": false,
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "471"
                  }
                }
              },
              "ETag": {
                "style": "simple",
                "explode": false,
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "W/\"1d7-8n/WZk1V5JEcCFkAefzBw4cGL10\""
                  }
                }
              },
              "Date": {
                "style": "simple",
                "explode": false,
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "Thu, 19 Aug 2021 03:56:36 GMT"
                  }
                }
              },
              "Connection": {
                "style": "simple",
                "explode": false,
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "keep-alive"
                  }
                }
              }
            },
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/ListOfSurveys1"
                },
                "example": {
                  "success": true,
                  "data": {
                    "questions": [
                      {
                        "description_image": "http://res.cloudinary.com/donifk2pu/image/upload/v1629341203/gqi4a8kvep8wlbshb2h6.jpg",
                        "_id": "611db933aab9cc6ab4912a2d",
                        "text": "updated text",
                        "createdAt": "2021-08-19T01:51:47.590+0000",
                        "__v": 0
                      },
                      {
                        "description_image": "#",
                        "_id": "611db934aab9cc6ab4912a31",
                        "text": "this is the second question",
                        "createdAt": "2021-08-19T01:51:48.419+0000",
                        "__v": 0
                      }
                    ],
                    "_id": "611db933aab9cc6ab4912a2b",
                    "name": "survey",
                    "description": "description"
                  }
                }
              }
            }
          }
        },
        "deprecated": false,
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        },
        "x-unitTests": [
          {
            "request": {
              "method": "GET",
              "uri": "/surveys/null"
            },
            "expectedResponse": {
              "x-allowExtraHeaders": true,
              "x-bodyMatchMode": "KEYS",
              "x-arrayOrderedMatching": false,
              "x-arrayCheckCount": false,
              "x-matchResponseSchema": true,
              "statusCode": "200",
              "headers": {
                "Content-Type": "application/json; charset=utf-8",
                "X-Powered-By": "Express",
                "Content-Length": "471",
                "ETag": "W/\"1d7-8n/WZk1V5JEcCFkAefzBw4cGL10\"",
                "Date": "Thu, 19 Aug 2021 03:56:36 GMT",
                "Connection": "keep-alive"
              },
              "body": "{\n\n  \"success\": true,\n\n  \"data\": {\n\n    \"questions\": [\n\n      {\n\n        \"description_image\": \"http://res.cloudinary.com/donifk2pu/image/upload/v1629341203/gqi4a8kvep8wlbshb2h6.jpg\",\n\n        \"_id\": \"611db933aab9cc6ab4912a2d\",\n\n        \"text\": \"updated text\",\n\n        \"createdAt\": \"2021-08-19T01:51:47.59Z\",\n\n        \"__v\": 0\n\n      },\n\n      {\n\n        \"description_image\": \"#\",\n\n        \"_id\": \"611db934aab9cc6ab4912a31\",\n\n        \"text\": \"this is the second question\",\n\n        \"createdAt\": \"2021-08-19T01:51:48.419Z\",\n\n        \"__v\": 0\n\n      }\n\n    ],\n\n    \"_id\": \"611db933aab9cc6ab4912a2b\",\n\n    \"name\": \"survey\",\n\n    \"description\": \"description\"\n\n  }\n\n}"
            },
            "x-testShouldPass": true,
            "x-testEnabled": true,
            "x-testName": "Get survey by id",
            "x-testDescription": "### Get survey by id\nreturn an object survey if found otherwise an empty object\n\n**protect route** : no\n**parameters** : id"
          }
        ]
      },
      "put": {
        "tags": [
          "Surveys"
        ],
        "summary": "Update a survey",
        "description": "## Update a survey given its id \naccepts as parameters the id of the survey and object containing attributes to update (only updates static attributes like name and description).\n\n**protected route** : yes\n\n**parameters** : id, survey",
        "operationId": "Updateasurvey",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "example": "null"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatedSurvey"
              },
              "example": {
                "name": "",
                "description": ""
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "X-Powered-By": {
                "style": "simple",
                "explode": false,
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "Express"
                  }
                }
              },
              "Content-Length": {
                "style": "simple",
                "explode": false,
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "76"
                  }
                }
              },
              "ETag": {
                "style": "simple",
                "explode": false,
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "W/\"4c-uyR46OKIkM4uLnRMPwlzwR8zL8M\""
                  }
                }
              },
              "Date": {
                "style": "simple",
                "explode": false,
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "Thu, 19 Aug 2021 04:00:09 GMT"
                  }
                }
              },
              "Connection": {
                "style": "simple",
                "explode": false,
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "keep-alive"
                  }
                }
              }
            },
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/Updateasurvey"
                },
                "example": {
                  "success": true,
                  "data": {
                    "_id": "611db933aab9cc6ab4912a2b",
                    "name": "updated name",
                    "description": "updated desc"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Updateasurvey2"
                },
                "example": {
                  "success": false,
                  "error": "Not authorized to access this route (admin only)"
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        },
        "x-unitTests": [
          {
            "request": {
              "method": "PUT",
              "uri": "/surveys/null",
              "headers": {
                "Content-Type": "application/json"
              },
              "body": "{\n\n  \"name\": \"\",\n\n  \"description\": \"\"\n\n}"
            },
            "expectedResponse": {
              "x-allowExtraHeaders": true,
              "x-bodyMatchMode": "KEYS",
              "x-arrayOrderedMatching": false,
              "x-arrayCheckCount": false,
              "x-matchResponseSchema": true,
              "statusCode": "200",
              "headers": {
                "Content-Type": "application/json; charset=utf-8",
                "X-Powered-By": "Express",
                "Content-Length": "76",
                "ETag": "W/\"4c-uyR46OKIkM4uLnRMPwlzwR8zL8M\"",
                "Date": "Thu, 19 Aug 2021 04:00:09 GMT",
                "Connection": "keep-alive"
              },
              "body": "{\n\n  \"success\": true,\n\n  \"data\": {\n\n    \"_id\": \"611db933aab9cc6ab4912a2b\",\n\n    \"name\": \"updated name\",\n\n    \"description\": \"updated desc\"\n\n  }\n\n}"
            },
            "x-testShouldPass": true,
            "x-testEnabled": true,
            "x-testName": "Update a survey",
            "x-testDescription": "## Update a survey given its id \naccepts as parameters the id of the survey and object containing attributes to update (only updates static attributes like name and description).\n\n**protected route** : yes\n\n**parameters** : id, survey"
          }
        ]
      },
      "delete": {
        "tags": [
          "Surveys"
        ],
        "summary": "Delete survey",
        "description": "### Delete survey by id\nDelete a survey with its id return error if the id is not found\n\n**protected route** : yes\n\n**parameters** : id",
        "operationId": "Deletesurvey",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "example": "null"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteSurveyMessage2"
                },
                "example": {
                  "success": true,
                  "data": {}
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deletesurvey1"
                },
                "example": {
                  "success": false,
                  "error": "Not authorized to access this route (admin only)"
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        },
        "x-unitTests": [
          {
            "request": {
              "method": "DELETE",
              "uri": "/surveys/null"
            },
            "expectedResponse": {
              "x-allowExtraHeaders": true,
              "x-bodyMatchMode": "KEYS",
              "x-arrayOrderedMatching": false,
              "x-arrayCheckCount": false,
              "x-matchResponseSchema": true,
              "statusCode": "200",
              "headers": {
                "Content-Type": "application/json"
              },
              "body": "{\n\n  \"success\": true,\n\n  \"data\": {}\n\n}"
            },
            "x-testShouldPass": true,
            "x-testEnabled": true,
            "x-testName": "Delete survey",
            "x-testDescription": "### Delete survey by id\nDelete a survey with its id return error if the id is not found\n\n**protected route** : yes\n\n**parameters** : id"
          }
        ]
      }
    },
...
    "/answers": {
      "get": {
        "tags": [
          "Answers"
        ],
        "summary": "Get all answers",
        "description": "### Get all answers\nreturn an array of all answers (generated from entries)\n\n**protected route** : yes\n\n**parameters** : none",
        "operationId": "Getallanswers",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Getallanswers2"
                },
                "example": {
                  "success": true
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Getallanswers1"
                },
                "example": {
                  "success": false,
                  "error": "Not authorized to access this route (admin only)"
                }
              }
            }
          },
          "default": {
            "content": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        },
        "x-unitTests": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ListOfSurveys1": {
        "title": "ListOfSurveys1",
        "required": [
          "data",
          "success"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/Survey1"
          }
        },
        "example": {
          "success": true,
          "data": {
            "questions": [
              {
                "description_image": "http://res.cloudinary.com/donifk2pu/image/upload/v1629341203/gqi4a8kvep8wlbshb2h6.jpg",
                "_id": "611db933aab9cc6ab4912a2d",
                "text": "updated text",
                "createdAt": "2021-08-19T01:51:47.590+0000",
                "__v": 0
              },
              {
                "description_image": "#",
                "_id": "611db934aab9cc6ab4912a31",
                "text": "this is the second question",
                "createdAt": "2021-08-19T01:51:48.419+0000",
                "__v": 0
              }
            ],
            "_id": "611db933aab9cc6ab4912a2b",
            "name": "survey",
            "description": "description"
          }
        }
      },
...
      "UpdatedSurvey": {
        "title": "UpdatedSurvey",
        "required": [
          "description",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "example": {
          "name": "",
          "description": ""
        }
      },
...
      "Updateasurvey": {
        "title": "Updateasurvey",
        "required": [
          "data",
          "success"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/UpdatedSurvey1"
          }
        },
        "example": {
          "success": true,
          "data": {
            "_id": "611db933aab9cc6ab4912a2b",
            "name": "updated name",
            "description": "updated desc"
          }
        }
      },
...
      "Updateasurvey2": {
        "title": "Updateasurvey2",
        "required": [
          "error",
          "success"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": "string"
          }
        },
        "example": {
          "success": false,
          "error": "Not authorized to access this route (admin only)"
        }
      },
      "Deletesurvey1": {
        "title": "Deletesurvey1",
        "required": [
          "error",
          "success"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": "string"
          }
        },
        "example": {
          "success": false,
          "error": "Not authorized to access this route (admin only)"
        }
      },
      "DeleteSurveyMessage2": {
        "title": "DeleteSurveyMessage2",
        "required": [
          "data",
          "success"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object"
          }
        },
        "example": {
          "success": true,
          "data": {}
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}